+++ packages/f-button/src/components/tests/Button.test.js
@@ -1,11 +1,11 @@
 import { shallowMount } from '@vue/test-utils';
-import VueButton from '../Button.vue';
+import FButton from '../Button.vue';
 
 describe('Button', () => {
     allure.feature('Button');
     it('should be defined', () => {
         const propsData = {};
-        const wrapper = shallowMount(VueButton, { propsData });
+        const wrapper = shallowMount(FButton, { propsData });
         expect(wrapper.exists()).toBe(true);
     });
 });