+++ packages/f-button/src/index.js
@@ -6,13 +6,13 @@
 
 
 // Import vue component
-import VueButton from '@/components/Button.vue';
+import FButton from '@/components/Button.vue';
 
 // Declare install function executed by Vue.use()
 export function install (Vue) {
     if (install.installed) return;
     install.installed = true;
-    Vue.component('VueButton', VueButton);
+    Vue.component('FButton', FButton);
 }
 
 // Create module definition for Vue.use()
@@ -32,4 +32,4 @@ if (GlobalVue) {
 }
 
 // To allow use as module (npm/webpack/etc.) export component
-export default VueButton;
+export default FButton;